+++ saleor/checkout/fetch.py
@@ -246,7 +246,7 @@ def fetch_checkout_info(
         checkout_info, shipping_address, lines, discounts, manager
     )
     valid_pick_up_points = get_valid_collection_points_for_checkout_info(
-        checkout_info, shipping_address, lines
+        shipping_address, lines, checkout_info
     )
     checkout_info.valid_shipping_methods = valid_shipping_methods
     checkout_info.valid_pick_up_points = valid_pick_up_points
@@ -297,15 +297,19 @@ def get_valid_shipping_method_list_for_checkout_info(
 
 
 def get_valid_collection_points_for_checkout_info(
-    checkout_info: "CheckoutInfo",
     shipping_address: Optional["Address"],
     lines: Iterable[CheckoutLineInfo],
+    checkout_info: CheckoutInfo,
 ):
     from .utils import get_valid_collection_points_for_checkout
 
-    country_code = shipping_address.country.code if shipping_address else None
+    if shipping_address:
+        country_code = shipping_address.country.code
+    else:
+        country_code = checkout_info.channel.default_country.code
+
     valid_collection_points = get_valid_collection_points_for_checkout(
-        lines, checkout_info, country_code=country_code, quantity_check=False
+        lines, country_code=country_code, quantity_check=False
     )
     return list(valid_collection_points)